content/canvas/test/test_toDataURL.default.html

<!DOCTYPE HTML>
<title>Canvas test: toDataURL.default</title>
<!-- Testing: toDataURL with no arguments returns a PNG -->
<script src="/MochiKit/packed.js"></script>
<script src="/tests/SimpleTest/SimpleTest.js"></script>
<link rel="stylesheet" href="/tests/SimpleTest/test.css">
<body>
<canvas id="c" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>
<script>

SimpleTest.waitForExplicitFinish();
MochiKit.DOM.addLoadEvent(function () {

var canvas = document.getElementById('c');
var ctx = canvas.getContext('2d');

var data = canvas.toDataURL();
ok(data =~ /^data:image\/png[;,]/, "data =~ /^data:image\\/png[;,]/");

SimpleTest.finish();

});
</script>
